+++ src/main/java/com/example/springboot/model/StudentRecord.java
@@ -30,7 +30,7 @@ public class StudentRecord {
     @Column(name = "first_name", nullable = false)
     private String firstName;
 
-    @Column(name = "middle_name", nullable = false)
+    @Column(name = "middle_name")
     private String middleName;
 
     @Column(name = "birthdate")